refactor: Move Round class to game package location
