+++ src/main/java/racinggame/game/RacingGame.java
@@ -1,7 +1,6 @@
 package racinggame.game;
 
 import racinggame.car.Car;
-import racinggame.car.Round;
 import racinggame.exception.RacingGameException;
 
 import java.util.List;
